Refactoring module struct into a separate file
